+++ explorer/src/lib/services/duskAPI.js
@@ -230,12 +230,26 @@ const duskAPI = {
 
   /**
    * @param {string} id
-   * @returns {Promise<Transaction>}
+   * @returns {Promise<Transaction | String | Error>}
    */
   getTransaction(id) {
     return gqlGet(gqlQueries.getTransactionQueryInfo(id))
       .then(getKey("tx"))
-      .then(transformTransaction);
+      .then((tx) => {
+        if (tx === null) {
+          return gqlGet(gqlQueries.getMempoolTx(id))
+            .then(getKey("tx"))
+            .then((mempoolTx) => {
+              if (mempoolTx) {
+                return "This transaction is currently in the mempool and has not yet been confirmed. The transaction details will be displayed after confirmation.";
+              } else {
+                throw new Error("Transaction not found");
+              }
+            });
+        } else {
+          return transformTransaction(tx);
+        }
+      });
   },
 
   /**
@@ -246,6 +260,16 @@ const duskAPI = {
     return gqlGet(gqlQueries.getTransactionDetailsQueryInfo(id)).then(
       getPath("tx.tx.json")
     );
+  },
+
+  /**
+   * @param {string} id
+   * @returns {Promise<Transaction>}
+   */
+  getTransactionFromMemPool(id) {
+    return gqlGet(gqlQueries.getMempoolTx(id))
+      .then(getKey("tx"))
+      .then(transformTransaction);
   },
 
   /**